refactoring alphanumeric name rules (#1569)
also refactoring common file name
