+++ src/cfnlint/rules/outputs/Name.py
@@ -2,10 +2,8 @@
 Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
-import re
 from cfnlint.rules import CloudFormationLintRule
-from cfnlint.rules import RuleMatch
-from cfnlint.helpers import REGEX_ALPHANUMERIC
+from cfnlint.rules.common import alphanumeric_name
 
 
 class Name(CloudFormationLintRule):
@@ -17,15 +15,4 @@ class Name(CloudFormationLintRule):
     tags = ['outputs']
 
     def match(self, cfn):
-        matches = []
-
-        outputs = cfn.template.get('Outputs', {})
-        for output_name, _ in outputs.items():
-            if not re.match(REGEX_ALPHANUMERIC, output_name):
-                message = 'Output {0} has invalid name.  Name has to be alphanumeric.'
-                matches.append(RuleMatch(
-                    ['Outputs', output_name],
-                    message.format(output_name)
-                ))
-
-        return matches
+        return alphanumeric_name(cfn, 'Outputs')
